requirements.txt
pillow
opencv-python
pynput
keyboard
PyAutoGUI
mss
numpy
imutils
pywin32
requests
pypiwin32